+++ test/text/test_cell.py
@@ -260,3 +260,26 @@ def test_cell_curfont_leak(tmp_path):  # issue #475
     pdf.cell(txt="XYZ012abc,-", new_x="LEFT", new_y="NEXT")
     pdf.cell(txt="3,7E-05", new_x="LEFT", new_y="NEXT")
     assert_pdf_equal(pdf, HERE / "cell_curfont_leak.pdf", tmp_path)
+
+
+# pylint: disable=protected-access
+def test_cell_lasth(tmp_path):  # issue #601
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("helvetica", size=18)
+    pdf.set_fill_color(255, 255, 0)
+    pdf.cell(w=100, txt="Hello world", fill=True)
+    print(pdf._lasth)  # prints: 6.35
+    pdf.ln()
+    assert pdf._lasth == 6.35, f"pdf._lasth ({pdf._lasth}) != 5.35"
+    pdf.set_fill_color(255, 0, 255)
+    pdf.cell(w=100, txt="Hello world", h=50, fill=True)
+    pdf.ln()
+    assert pdf._lasth == 50, f"pdf._lasth ({pdf._lasth}) != 50 after cell(h=50)"
+    pdf.set_fill_color(0, 255, 255)
+    pdf.cell(w=100, txt="Hello world", fill=True)
+    pdf.cell(w=100, txt="")
+    pdf.ln()
+    assert pdf._lasth == 6.35, f"pdf._lasth ({pdf._lasth}) != 5.35 after empty cell"
+    pdf.cell(w=100, txt="Hello world", border=True)
+    assert_pdf_equal(pdf, HERE / "cell_lasth.pdf", tmp_path)